servlet:upload file

+++ server/src/main/java/com/netnovelreader/servlet/rule/DeleteSitePreference.kt
@@ -1,6 +1,7 @@
 package com.netnovelreader.servlet.rule
 
 import com.netnovelreader.service.SitePreferenceService
+import com.netnovelreader.servlet.BaseAuthorityServlet
 import java.io.IOException
 import javax.servlet.ServletException
 import javax.servlet.annotation.WebServlet
@@ -12,14 +13,16 @@ class DeleteSitePreference : BaseAuthorityServlet() {
     @Throws(ServletException::class, IOException::class)
     override fun doGet(req: HttpServletRequest, resp: HttpServletResponse) {
         super.doGet(req, resp)
-
+        if (role != 100) {
+            resp.sendRedirect("redirect.jsp")
+            return
+        }
         SitePreferenceService().deletePreference(req.getParameter("hostname"))
         resp.sendRedirect("index.jsp")
     }
 
     @Throws(ServletException::class, IOException::class)
     override fun doPost(req: HttpServletRequest, resp: HttpServletResponse) {
-        super.doPost(req, resp)
         doGet(req, resp)
     }
 }
